Implemented dijkstra

+++ names.h
@@ -4,7 +4,7 @@
 
 #include <vector>
 
-using node_name_t = unsigned int;
+using node_name_t = size_t;
 using node_data_t = bool; // maybe temp
 using weight_t = double;
 using graph_t = graph::Graph<node_name_t, node_data_t, weight_t>;
